Refactor retriever into module with subdir
