+++ hc-dna/zomes/perspective_diff_sync/src/retriever/holochain.rs
@@ -1,0 +1,20 @@
+use perspective_diff_sync_integrity::{PerspectiveDiffEntryReference};
+use std::collections::BTreeMap;
+use hdk::prelude::*;
+use std::sync::Mutex;
+
+pub struct HolochainRetreiver;
+
+impl PerspectiveDiffRetreiver for HolochainRetreiver {
+    fn get(hash: Hash) -> SocialContextResult<PerspectiveDiffEntryReference> {
+        get(hash, GetOptions::latest())?
+            .ok_or(SocialContextError::InternalError(
+                "HolochainRetreiver: Could not find entry while populating search",
+            ))?
+            .entry()
+            .to_app_option::<PerspectiveDiffEntryReference>()?
+            .ok_or(SocialContextError::InternalError(
+                "Expected element to contain app entry data",
+            ))
+    }
+}
